Support list of devices in config file
Permit multiple instances of a sensor type to be specified in the
config file as a list, e.g. two Raritan units. This needs to be done
because in JSON the keys are expected to be globally unique and in case
of same key appearing more than once, the last one takes effect. This
change overcomes the nasty implicit assumption that LabSense made about
there being only instance of a sensor of a given type.